feat: theme changes avoid relayout if not necessary
When layout needs to be recalculated, we have to rebuild the entire app.
Right now, because all views are kept in memory (recents, games, apps,
settings), this takes long to load.

For now, until we split out the views, we avoid relayout when not
necessary. The most common use case is changing colors, which does not
actually require relayout.